+++ pom.xml
@@ -35,7 +35,7 @@
 							</descriptorRefs>
 							<archive>
 								<manifest>
-									<mainClass>tk.yabl.main.Main</mainClass>
+									<mainClass>xyz.yabl.main.Main</mainClass>
 								</manifest>
 							</archive>
 						</configuration>
