+++ src/logback.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <configuration>
+
 	<appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
 		<layout class="ch.qos.logback.classic.PatternLayout">
 			<Pattern>
@@ -8,17 +9,21 @@
 		</layout>
 	</appender>
 	<appender name="FILE" class="ch.qos.logback.core.FileAppender">
-		<file>log.log</file>
-		<append>false</append>
-		<!-- set immediateFlush to false for much higher logging throughput -->
-		<encoder>
+	    <file>log.log</file>
+	    <append>false</append>
+	    <!-- set immediateFlush to false for much higher logging throughput -->
+	    
+	    <encoder>
 			<Pattern>
 				%d{yyyy-MM-dd HH:mm:ss} [%thread] %-5level %logger{36} - %msg%n
 			</Pattern>
 		</encoder>
-	</appender>
+  </appender>
+
 	<root level="info">
 		<appender-ref ref="STDOUT" />
 		<appender-ref ref="FILE" />
 	</root>
+	
+
 </configuration>
