+++ www/404.html
@@ -1,18 +1,17 @@
 <!DOCTYPE html>
 <html lang="en">
-
 	<head>
 		<title>Yet Another Bot List</title>
-		<meta charset="utf-8" />
+		<meta charset="utf-8"/>
 		<link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet">
 		<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.1/css/all.css" integrity="sha384-gfdkjb5BdAXd+lj+gudLWI+BXq4IuLW5IT+brZEZsLFm++aCMlF1V92rMkPaX4PP" crossorigin="anonymous">
 		<link rel="stylesheet" type="text/css" href="/style.css">
 	</head>
-
 	<body>
 		<div class="navbar">
 			<div class="listitem home" onclick="home()">
-				<div class="hover"></div>
+				<div class="hover">
+				</div>
 				<p class="navtext text">YABL</p>
 			</div>
 			<div class="rightnav">
@@ -32,81 +31,74 @@
 					<p class="navtext text"><i class="fas fa-plus-circle"></i></p>
 				</a>
 				<a href="/verify" class="listitem verifynav">
+						<div class="hover"></div>
+						<p class="navtext text"><i class="fas fa-user-check"></i></p>
+					</div>
+				<a href="/dashboard" class="listitem profile">
 					<div class="hover"></div>
-					<p class="navtext text"><i class="fas fa-user-check"></i></p>
+					<p class="navtext text"><i class="fas fa-user-circle"></i></p>
 				</a>
 			</div>
-			<a href="/dashboard" class="listitem profile">
-				<div class="hover"></div>
-				<p class="navtext text"><i class="fas fa-user-circle"></i></p>
-			</a>
 		</div>
 		<div class="pagecontent errorcontent">
-			<p class="text header">404</p>
-			<p class="text subheader">Not Found</p>
-			<br />
-			<p class="text">
-				The resource you're trying to access was not found by the server.<br />
-				If you believe the resource actually does exist, please contact us on <a href="https://discord.gg/92CazE">the discord server.</a>
-			</p>
+			<p class="text header">
+				404
+            </p>
+            <p class="text subheader">
+                Not Found
+            </p>
+            <br/>
+            <p class="text">
+                The resource you're trying to access was not found by the server.<br/>
+                If you believe the resource actually does exist, please contact us on <a href="https://discord.gg/92CazE">the discord server.</a>
+            </p>
+      </div>
 		</div>
 		<div class="footer">
 			<p class="footitem text">&#169; YABL team 2018-19, built with java and html.</p>
 			<p class="footitem text"><a href="/terms">Terms</a>|<a href="https://github.com/yabl-team">Github</a>|<a href="/api">API</a></p>
 		</div>
 		<script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-		<script language="javascript">
-			var baseUrl = window.location.protocol + "//" + window.location.hostname;
-			if (window.localStorage.getItem("token") != null) {
-				document.querySelector(".listitem.login").style.display = "none";
-			} else {
-				document.querySelector(".listitem.logout").style.display = "none";
-				document.querySelector(".listitem.addbot").style.display = "none";
-				document.querySelector(".listitem.profile").style.display = "none";
+		<script>
+			var baseUrl = window.location.protocol +"//"+ window.location.hostname
+			if(window.localStorage.getItem("token") != null){
+			  document.querySelector(".listitem.login").style.display = "none"
+			} else{
+			  document.querySelector(".listitem.logout").style.display = "none"
+			  document.querySelector(".listitem.addbot").style.display = "none"
+			  document.querySelector(".listitem.profile").style.display = "none"
 			}
-
-			function home() {
-				window.location.href = baseUrl + "/";
+            function home(){
+                window.location.href = baseUrl + "/";
+            }
+			
+			function logout(){
+			  window.localStorage.removeItem("token");
+			  window.location.reload();
 			}
-
-			function logout() {
-				window.localStorage.removeItem("token");
-				window.location.reload();
-			}
-
+			
+			
+			
 			document.querySelector(".searchbox").addEventListener("keyup", function(event) {
-				event.preventDefault();
-				if (event.keyCode === 13) {
-					window.location.href = baseUrl + "/search?query=" + encodeURI(document.querySelector(".searchbox").value);
-				}
+			  event.preventDefault();
+			  if (event.keyCode === 13) {
+			    window.location.href = baseUrl + "/search?query="+encodeURI(document.querySelector(".searchbox").value);
+			  }
 			});
-
-			var instance = axios.create({
-				validateStatus: a => {
-					return true;
-				}
-			});
-
+			var instance = axios.create({validateStatus: a => {return true;}});
 			var whoami = null;
-			if (localStorage.getItem("token") !== null) {
-				instance.get(baseUrl + "/api/whoami", {
-					"headers": {
-						"Content-Type": "application/json",
-						"Authorization": localStorage.getItem("token")
+            if(localStorage.getItem("token") !== null){
+                instance.get(baseUrl+"/api/whoami",{"headers":{"Content-Type":"application/json","Authorization":localStorage.getItem("token")}}).then(function(response){whoami = response.data
+                    if(response.status == "401") {
+					localStorage.removeItem("token")
+					window.location.href = baseUrl + "/login"
 					}
-				}).then(function(response) {
-					whoami = response.data;
-					if (response.status == "401") {
-						localStorage.removeItem("token");
-						window.location.href = baseUrl + "/login";
-					}
-					if (!whoami.admin) document.querySelector(".listitem.verifynav").style.display = "none";
-				}).catch();
-			} else {
-				document.querySelector(".listitem.verifynav").style.display = "none";
-			}
-
+					if(!whoami.admin) document.querySelector(".listitem.verifynav").style.display = "none"
+                }).catch();
+            } else{
+                document.querySelector(".listitem.verifynav").style.display = "none"
+            }
+            
 		</script>
 	</body>
-
 </html>
