+++ www/flagged.html
@@ -1,79 +1,77 @@
 <!DOCTYPE html>
 <html lang="en">
-
-	<head>
-		<title>Yet Another Bot List</title>
-		<meta charset="utf-8" />
-		<meta name="viewport" content="width=device-width,initial-scale=1,maximum-scale=1,user-scalable=no">
-		<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.1/css/all.css" integrity="sha384-gfdkjb5BdAXd+lj+gudLWI+BXq4IuLW5IT+brZEZsLFm++aCMlF1V92rMkPaX4PP" crossorigin="anonymous">
-		<link rel="stylesheet" type="text/css" href="/style.css">
-	</head>
-
-	<body>
-		<div class="navbar">
-			<div class="listitem home" onclick="home()">
-				<div class="hover"></div>
-				<p class="navtext text">YABL</p>
-			</div>
-			<div class="rightnav">
-				<div class="listitem search">
-					<input class="searchbox inputbox" placeholder="Search" onenter="search">
-				</div>
-				<a href="/login" class="listitem login">
-					<div class="hover"></div>
-					<p class="navtext text"><i class="fas fa-sign-in-alt"></i></p>
-				</a>
-				<div class="listitem logout" onclick="logout()">
-					<div class="hover"></div>
-					<p class="navtext text"><i class="fas fa-sign-out-alt"></i></p>
-				</div>
-				<a href="/add" class="listitem addbot">
-					<div class="hover"></div>
-					<p class="navtext text"><i class="fas fa-plus-circle"></i></p>
-				</a>
-				<a href="/verify" class="listitem verifynav">
-					<div class="hover"></div>
-					<p class="navtext text"><i class="fas fa-user-check"></i></p>
-				</a>
-			</div>
-			<a href="/dashboard" class="listitem profile">
-				<div class="hover"></div>
-				<p class="navtext text"><i class="fas fa-user-circle"></i></p>
-			</a>
-		</div>
-		<div class="pagecontent"></div>
-		<div class="footer">
-			<p class="footitem text">&#169; YABL team 2018, built with java and html.</p>
-			<p class="footitem text"><a href="/terms">Terms</a>|<a href="https://github.com/yabl-team">Github</a></p>
-		</div>
-		<script language="javascript">
-			var baseUrl = window.location.protocol + "//" + window.location.hostname;
-			if (window.localStorage.getItem("token") != null) {
-				document.querySelector(".listitem.login").style.display = "none";
-			} else {
-				document.querySelector(".listitem.logout").style.display = "none";
-				document.querySelector(".listitem.addbot").style.display = "none";
-				document.querySelector(".listitem.profile").style.display = "none";
-			}
-
-			function logout() {
-				window.localStorage.removeItem("token");
-				window.location.reload();
-			}
-
-			function home() {
-				window.location.href = baseUrl;
-			}
-
-			document.querySelector(".searchbox").addEventListener("keyup", function(event) {
-				event.preventDefault();
-				if (event.keyCode === 13) {
-					window.location.href = baseUrl + "/search?query=" + encodeURI(document.querySelector(".searchbox").value);
-				}
-			});
-
-		</script>
-		<script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-	</body>
-
+    <head>
+        <title>Yet Another Bot List</title>
+        <meta charset="utf-8"/>
+        <meta name="viewport" content="width=device-width,initial-scale=1,maximum-scale=1,user-scalable=no">
+        <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.6.1/css/all.css" integrity="sha384-gfdkjb5BdAXd+lj+gudLWI+BXq4IuLW5IT+brZEZsLFm++aCMlF1V92rMkPaX4PP" crossorigin="anonymous">
+        <link rel="stylesheet" type="text/css" href="/style.css">
+    </head>
+    <body>
+        <div class="navbar">
+            <div class="listitem home" onclick="home()">
+              <div class="hover"></div>
+              <p class="navtext text">YABL</p>
+            </div>
+            <div class="rightnav">
+                <div class="listitem search">
+                    <input class="searchbox inputbox" placeholder="Search" onenter="search"></div>
+                </div>
+                <a href="/login" class="listitem login">
+                  <div class="hover"></div>
+                  <p class="navtext text"><i class="fas fa-sign-in-alt"></i></p>
+                </a>
+                <div class="listitem logout" onclick="logout()">
+                  <div class="hover"></div>
+                  <p class="navtext text"><i class="fas fa-sign-out-alt"></i></p>
+                </div>
+                <a href="/add" class="listitem addbot">
+                  <div class="hover"></div>
+                  <p class="navtext text"><i class="fas fa-plus-circle"></i></p>
+                </a>
+                <a href="/verify" class="listitem verifynav">
+                    <div class="hover"></div>
+                    <p class="navtext text"><i class="fas fa-user-check"></i></p>
+                  </div>
+                <a href="/dashboard" class="listitem profile">
+                  <div class="hover"></div>
+                  <p class="navtext text"><i class="fas fa-user-circle"></i></p>
+                </a>
+            </div>
+        </div>
+        <div class="pagecontent">
+          
+        </div>
+        <div class="footer">
+            <p class="footitem text">&#169; YABL team 2018, built with java and html.</p>
+            <p class="footitem text"><a href="/terms">Terms</a>|<a href="https://github.com/yabl-team">Github</a></p>
+        </div>
+        <script>
+            var baseUrl = window.location.protocol +"//"+ window.location.hostname
+            if(window.localStorage.getItem("token") != null){
+              document.querySelector(".listitem.login").style.display = "none"
+            } else{
+              document.querySelector(".listitem.logout").style.display = "none"
+              document.querySelector(".listitem.addbot").style.display = "none"
+              document.querySelector(".listitem.profile").style.display = "none"
+            }
+            
+            function logout(){
+              window.localStorage.removeItem("token");
+              window.location.reload();
+            }
+            
+            
+            function home(){
+                window.location.href = baseUrl;
+            }
+            document.querySelector(".searchbox").addEventListener("keyup", function(event) {
+              event.preventDefault();
+              if (event.keyCode === 13) {
+                window.location.href = baseUrl + "/search?query="+encodeURI(document.querySelector(".searchbox").value);
+              }
+            });
+        </script>
+        <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
+    </body>
 </html>
